create data_processing.py
